Migrate httpdispatcher to caddy2
